Check if a user verified after each login
