+++ lib/binding/email_verification_binding.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+import 'package:satorio/controller/email_verification_controller.dart';
+
+class EmailVerificationBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<EmailVerificationController>(
+      () => EmailVerificationController(),
+    );
+  }
+}
