+++ lib/controller/login_controller.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
 import 'package:satorio/binding/create_account_binding.dart';
+import 'package:satorio/binding/email_verification_binding.dart';
 import 'package:satorio/binding/main_binding.dart';
 import 'package:satorio/binding/password_recovery_binding.dart';
 import 'package:satorio/controller/mixin/validation_mixin.dart';
 import 'package:satorio/domain/repositories/sator_repository.dart';
 import 'package:satorio/ui/page_widget/create_account_page.dart';
+import 'package:satorio/ui/page_widget/email_verification_page.dart';
 import 'package:satorio/ui/page_widget/forgot_password_page.dart';
 import 'package:satorio/ui/page_widget/main_page.dart';
 
@@ -30,8 +32,20 @@ class LoginController extends GetxController with ValidationMixin {
         .signIn(emailController.text, passwordController.text)
         .then((isSuccess) {
       if (isSuccess) {
-        Get.offAll(() => MainPage(), binding: MainBinding());
+        _checkIsVerified();
       }
     }).catchError((value) => handleValidationException(value));
+  }
+
+  void _checkIsVerified() {
+    _satorioRepository.isVerified().then((isVerified) {
+      if (isVerified)
+        Get.offAll(() => MainPage(), binding: MainBinding());
+      else
+        Get.to(
+          () => EmailVerificationPage(),
+          binding: EmailVerificationBinding(),
+        );
+    });
   }
 }
