+++ lib/controller/splash_controller.dart
@@ -1,7 +1,9 @@
 import 'package:get/get.dart';
+import 'package:satorio/binding/email_verification_binding.dart';
 import 'package:satorio/binding/main_binding.dart';
 import 'package:satorio/binding/onboarding_binding.dart';
 import 'package:satorio/domain/repositories/sator_repository.dart';
+import 'package:satorio/ui/page_widget/email_verification_page.dart';
 import 'package:satorio/ui/page_widget/main_page.dart';
 import 'package:satorio/ui/page_widget/onboardinga_page.dart';
 
@@ -20,11 +22,23 @@ class SplashController extends GetxController {
     Future.delayed(Duration(milliseconds: 1000), () {
       _satorioRepository.isTokenValid().then((isTokenValid) {
         if (isTokenValid) {
-          Get.offAll(() => MainPage(), binding: MainBinding());
+          _checkIsVerified();
         } else {
           Get.off(() => OnBoardingPage(), binding: OnBoardingBinding());
         }
       });
+    });
+  }
+
+  void _checkIsVerified() {
+    _satorioRepository.isVerified().then((isVerified) {
+      if (isVerified)
+        Get.offAll(() => MainPage(), binding: MainBinding());
+      else
+        Get.off(
+          () => EmailVerificationPage(),
+          binding: EmailVerificationBinding(),
+        );
     });
   }
 }
