+++ lib/ui/page_widget/forgot_password_reset.dart
@@ -23,9 +23,9 @@ class ForgotPasswordResetPage extends GetView<PasswordRecoveryController> {
           child: InkWell(
             onTap: () => controller.back(),
             child: Icon(
-              Icons.arrow_back_ios,
+              Icons.chevron_left_rounded,
               color: SatorioColor.darkAccent,
-              size: 18,
+              size: 32,
             ),
           ),
         ),
